==> 2_Tratamento_de_dados.ipynb <==
"""
    Atividade introdutória no Tratamento de dados em python

👨‍💻 Autor: CLEITON HENTGES
📅 Data: [29/08/2025]
💻 Linguagem: Python 3.x
"""

import pandas as pd

df = pd.read_csv('/content/clientes.csv')

# Verificar os primeiros registros
print(df.head().to_string(),'\n')

print(df.tail().to_string(),'\n')

# Verificar qtd de linhas e colunas
print('Qtd: ', df.shape,'\n')

# Verificar tipos de dados
print('Tipos:\n', df.dtypes,'\n')

# Checar valores nulos
print('Valores nulos:\n', df.isnull().sum(),'\n')

# Remover dados
df.drop(labels='pais', axis=1, inplace=True)   # Coluna
df.drop(labels=2, axis=0, inplace=True)        # Linha

# Normalizar campos de texto
df['nome'] = df['nome'].str.title()
df['endereco'] = df['endereco'].str.lower()
df['estado'] = df['estado'].str.strip().str.upper()

# Converter tipos de dados
df['idade'] = df['idade'].astype(int)

# Tratar valores nulos (ausentes)
df_fillna = df.fillna(0)                       # Substituir valores nulos por 0
df_dropna = df.dropna()                        # Remover registros com valores nulos
df_dropna4 = df.dropna(thresh=4)               # Manter registros com no mínimo 4 valores não nulos
df = df.dropna(subset=['cpf'])                 # Remover registros com CPF nulo

print('Valores nulos:\n', df.isnull().sum())

print('Qtd de registros nulos com fillna:', df_fillna.isnull().sum().sum())
print('Qtd de registros nulos com dropna:', df_dropna.isnull().sum().sum())
print('Qtd de registros nulos com dropna4:', df_dropna4.isnull().sum().sum())
print('Qtd de registros nulos com CPF:', df.isnull().sum().sum())

# Preencher valores específicos em colunas
df.fillna(value={'estado': 'Desconhecido'}, inplace=True)
df['endereco'] = df['endereco'].fillna('Endereço não informado')
df['idade_corrigida'] = df['idade'].fillna(df['idade'].mean())

# Tratar formato de dados
df['data_corrigida'] = pd.to_datetime(df['data'], format='%d/%m/%Y', errors='coerce')

# Tratar dados duplicados
print('Qtd registros atual:', df.shape[0])

df.duplicated()  # Identificar duplicados
df.drop_duplicates(subset='cpf', inplace=True)  # Remover duplicados com base no CPF

print('Qtd registros removendo as duplicadas:', len(df),'n')

print('Dados Limpos:\n', df)

# Salvar dataframe
df['data'] = df['data_corrigida']
df['idade'] = df['idade_corrigida']

df_salvar = df[['nome', 'cpf', 'idade', 'data', 'endereco', 'estado']]
df_salvar.to_csv('clientes_limpeza.csv', index=False)

print('\nNovo DataFrame:\n', pd.read_csv('clientes_limpeza.csv'))

# Função para calcular o cubo de um número
def eleva_cubo(x):
    return x ** 3

# Expressão de lambda para calcular o cubo de um número
eleva_cubo_lambda = lambda x: x ** 3

print(eleva_cubo(2))
print(eleva_cubo_lambda(2))

df = pd.DataFrame({'numeros': [1, 2, 3, 4, 5, 10]})

df['cubo_funcao'] = df['numeros'].apply(eleva_cubo)
df['cubo_lambda'] = df['numeros'].apply(lambda x: x ** 3)

print(df)
